# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrainCoach API",
    "version": "v1"
  },
  "paths": {
...
    "/api/integrations/{provider}/authorize-url": {
      "get": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationUrlDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationUrlDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationUrlDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/callback": {
      "post": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/connect-demo": {
      "post": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}": {
      "delete": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/integrations/{provider}/sync": {
      "post": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/integrations/{provider}/sync-history": {
      "get": {
        "tags": [
          "IntegrationConnections"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IntegrationProviderType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SynchronizationRunDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SynchronizationRunDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SynchronizationRunDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/athletes/{athleteUserId}/personal-records": {
      "get": {
        "tags": [
          "PersonalRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PersonalRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalRecordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalRecordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/personal-records/{recordId}": {
      "put": {
        "tags": [
          "PersonalRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalRecordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalRecordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecordDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PersonalRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/athletes/{athleteUserId}/recovery": {
      "get": {
        "tags": [
          "RecoveryMetrics"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryMetricDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryMetricDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryMetricDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RecoveryMetrics"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRecoveryMetricRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRecoveryMetricRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRecoveryMetricRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryMetricDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryMetricDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryMetricDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorizationUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreatePersonalRecordRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "distanceLabel": {
            "type": "string",
            "nullable": true
          },
          "timeSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "achievedDate": {
            "type": "string",
            "format": "date"
          },
          "raceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataSource": {
        "enum": [
          "Manual",
          "FileImport",
          "Strava",
          "GarminDemoProvider",
          "MySasyDemoProvider"
        ],
        "type": "string"
      },
...
      "IntegrationConnectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProviderType"
          },
          "status": {
            "$ref": "#/components/schemas/IntegrationConnectionStatus"
          },
          "externalAccountId": {
            "type": "string",
            "nullable": true
          },
          "connectedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSyncedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegrationConnectionStatus": {
        "enum": [
          "NotConnected",
          "Connected",
          "Error",
          "Revoked"
        ],
        "type": "string"
      },
      "IntegrationProviderType": {
        "enum": [
          "Strava",
          "GarminDemoProvider",
          "MySasyDemoProvider"
        ],
        "type": "string"
      },
...
      "OAuthCallbackRequest": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PersonalRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "distanceLabel": {
            "type": "string",
            "nullable": true
          },
          "timeSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "achievedDate": {
            "type": "string",
            "format": "date"
          },
          "raceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RecoveryMetricDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "restingHeartRateBpm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "readinessScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "stressScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          }
        },
        "additionalProperties": false
      },
...
      "SportType": {
        "enum": [
          "Running",
          "Cycling",
          "Swimming",
          "Strength",
          "CrossTraining",
          "Rest",
          "Other"
        ],
        "type": "string"
      },
...
      "SyncRunStatus": {
        "enum": [
          "Pending",
          "Running",
          "Succeeded",
          "Failed"
        ],
        "type": "string"
      },
      "SynchronizationRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProviderType"
          },
          "status": {
            "$ref": "#/components/schemas/SyncRunStatus"
          },
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "itemsFetched": {
            "type": "integer",
            "format": "int32"
          },
          "itemsCreated": {
            "type": "integer",
            "format": "int32"
          },
          "itemsSkippedDuplicate": {
            "type": "integer",
            "format": "int32"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePersonalRecordRequest": {
        "type": "object",
        "properties": {
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "distanceLabel": {
            "type": "string",
            "nullable": true
          },
          "timeSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "achievedDate": {
            "type": "string",
            "format": "date"
          },
          "raceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpsertRecoveryMetricRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "restingHeartRateBpm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "readinessScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "stressScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Zadejte JWT access token (bez prefixu \"Bearer \").",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}